@@ -11,7 +11,10 @@
 		}
 	],
 	"require": {
-		"php": ">=5.4.0"
+		"php": ">=5.6.0"
+	},
+	"suggest": {
+		"ext-redis": "Improved performance for communicating with redis"
 	},
 	"autoload": {
 		"classmap": [
