@@ -4,5 +4,5 @@ require_once dirname(__FILE__).'/CredisSentinelTest.php';
 
 class CredisStandaloneSentinelTest extends CredisSentinelTest
 {
-    protected $useStandalone = TRUE;
+    protected $useStandalone = true;
 }
