@@ -8,25 +8,25 @@ class CredisTest extends CredisTestCommon
     /** @var Credis_Client */
     protected $credis;
 
-    protected function setUp()
+    protected function setUpInternal()
     {
-        parent::setUp();
+        parent::setUpInternal();
         $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout']);
-        if($this->useStandalone) {
+        if ($this->useStandalone) {
             $this->credis->forceStandalone();
         }
         $this->credis->flushDb();
     }
-    protected function tearDown()
+    protected function tearDownInternal()
     {
-        if($this->credis) {
+        if ($this->credis) {
             $this->credis->close();
-            $this->credis = NULL;
+            $this->credis = null;
         }
     }
     public function testFlush()
     {
-        $this->credis->set('foo','FOO');
+        $this->credis->set('foo', 'FOO');
         $this->assertTrue($this->credis->flushDb());
         $this->assertFalse($this->credis->get('foo'));
     }
@@ -61,7 +61,7 @@ class CredisTest extends CredisTestCommon
     public function testScalars()
     {
         // Basic get/set
-        $this->credis->set('foo','FOO');
+        $this->credis->set('foo', 'FOO');
         $this->assertEquals('FOO', $this->credis->get('foo'));
         $this->assertFalse($this->credis->get('nil'));
 
@@ -70,12 +70,12 @@ class CredisTest extends CredisTestCommon
         $this->assertEquals($this->credis->exists('nil'), 0);
 
         // Empty string
-        $this->credis->set('empty','');
+        $this->credis->set('empty', '');
         $this->assertEquals('', $this->credis->get('empty'));
 
         // UTF-8 characters
         $utf8str = str_repeat("quarter: ¼, micro: µ, thorn: Þ, ", 500);
-        $this->credis->set('utf8',$utf8str);
+        $this->credis->set('utf8', $utf8str);
         $this->assertEquals($utf8str, $this->credis->get('utf8'));
 
         // Array
@@ -84,16 +84,26 @@ class CredisTest extends CredisTestCommon
         $this->assertTrue(in_array('FOO', $mGet));
         $this->assertTrue(in_array('BAR', $mGet));
         $this->assertTrue(in_array('', $mGet));
+        $mGet = $this->credis->mGet(array('foo'));
+        $this->assertTrue(in_array('FOO', $mGet));
+        $mGet = $this->credis->mGet(array());
+        $this->assertTrue($mGet === false);
 
         // Non-array
-        $mGet = $this->credis->mGet('foo','bar');
+        $mGet = $this->credis->mGet('foo', 'bar');
         $this->assertTrue(in_array('FOO', $mGet));
         $this->assertTrue(in_array('BAR', $mGet));
+        $mGet = $this->credis->mGet('foo');
+        $this->assertTrue(in_array('FOO', $mGet));
 
-        // Delete strings, null response
-        $this->assertEquals(2, $this->credis->del('foo','bar'));
+        // Delete strings and check they are deleted
+        $this->assertEquals(2, $this->credis->del('foo', 'bar'));
         $this->assertFalse($this->credis->get('foo'));
         $this->assertFalse($this->credis->get('bar'));
+
+        // Unlink strings using array args
+        $this->assertTrue($this->credis->mSet(['foo' => 'FOO', 'bar' => 'BAR']));
+        $this->assertEquals(2, $this->credis->unlink(['foo', 'bar']));
 
         // Long string
         $longString = str_repeat(md5('asd'), 4096); // 128k (redis.h REDIS_INLINE_MAX_SIZE = 64k)
@@ -141,9 +151,9 @@ class CredisTest extends CredisTestCommon
         $this->assertEquals(10, $range['And']);
 
         // withscores-option is off
-        $range = $this->credis->zRange('myset', 0, 4, array('withscores'));
-        $this->assertEquals(4, count($range));
-        $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
+        // $range = $this->credis->zRange('myset', 0, 4, array('withscores'));
+        // $this->assertEquals(4, count($range));
+        // $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
 
         $range = $this->credis->zRange('myset', 0, 4, array('withscores' => false));
         $this->assertEquals(4, count($range));
@@ -191,9 +201,9 @@ class CredisTest extends CredisTestCommon
         $this->assertEquals(11, $range['Goodbye']);
 
         // withscores-option is off
-        $range = $this->credis->zRangeByScore('myset', '-inf', '+inf', array('withscores'));
-        $this->assertEquals(4, count($range));
-        $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
+        // $range = $this->credis->zRangeByScore('myset', '-inf', '+inf', array('withscores'));
+        // $this->assertEquals(4, count($range));
+        // $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
 
         $range = $this->credis->zRangeByScore('myset', '-inf', '+inf', array('withscores' => false));
         $this->assertEquals(4, count($range));
@@ -214,7 +224,7 @@ class CredisTest extends CredisTestCommon
         $this->assertTrue(array_key_exists('And', $range));
         $this->assertEquals(10, $range['And']);
 
-        $range = $this->credis->zRevRangeByScore('myset', '+inf',10, array('withscores' => true));
+        $range = $this->credis->zRevRangeByScore('myset', '+inf', 10, array('withscores' => true));
         $this->assertEquals(2, count($range));
         $this->assertTrue(array_key_exists('And', $range));
         $this->assertEquals(10, $range['And']);
@@ -222,9 +232,9 @@ class CredisTest extends CredisTestCommon
         $this->assertEquals(11, $range['Goodbye']);
 
         // withscores-option is off
-        $range = $this->credis->zRevRangeByScore('myset', '+inf', '-inf', array('withscores'));
-        $this->assertEquals(4, count($range));
-        $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
+        // $range = $this->credis->zRevRangeByScore('myset', '+inf', '-inf', array('withscores'));
+        // $this->assertEquals(4, count($range));
+        // $this->assertEquals(range(0, 3), array_keys($range)); // expecting numeric array without scores
 
         $range = $this->credis->zRevRangeByScore('myset', '+inf', '-inf', array('withscores' => false));
         $this->assertEquals(4, count($range));
@@ -267,12 +277,12 @@ class CredisTest extends CredisTestCommon
 
     public function testHashes()
     {
-        $this->assertEquals(1, $this->credis->hSet('hash','field1','foo'));
-        $this->assertEquals(0, $this->credis->hSet('hash','field1','foo'));
-        $this->assertEquals('foo', $this->credis->hGet('hash','field1'));
-        $this->assertEquals(NULL, $this->credis->hGet('hash','x'));
+        $this->assertEquals(1, $this->credis->hSet('hash', 'field1', 'foo'));
+        $this->assertEquals(0, $this->credis->hSet('hash', 'field1', 'foo'));
+        $this->assertEquals('foo', $this->credis->hGet('hash', 'field1'));
+        $this->assertEquals(null, $this->credis->hGet('hash', 'x'));
         $this->assertTrue($this->credis->hMSet('hash', array('field2' => 'Hello', 'field3' => 'World')));
-        $this->assertEquals(array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => FALSE), $this->credis->hMGet('hash', array('field1','field2','nilfield')));
+        $this->assertEquals(array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => false), $this->credis->hMGet('hash', array('field1','field2','nilfield')));
         $this->assertEquals(array(), $this->credis->hGetAll('nohash'));
         $this->assertEquals(array('field1' => 'foo', 'field2' => 'Hello', 'field3' => 'World'), $this->credis->hGetAll('hash'));
 
@@ -290,11 +300,11 @@ class CredisTest extends CredisTestCommon
 
         $this->credis->pipeline();
         $this->assertTrue($this->credis === $this->credis->hMGet('hash', array('field1','field2','nilfield')));
-        $this->assertEquals(array(0 => array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => FALSE)), $this->credis->exec());
+        $this->assertEquals(array(0 => array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => false)), $this->credis->exec());
 
         $this->credis->pipeline()->multi();
         $this->assertTrue($this->credis === $this->credis->hMGet('hash', array('field1','field2','nilfield')));
-        $this->assertEquals(array(0 => array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => FALSE)), $this->credis->exec());
+        $this->assertEquals(array(0 => array('field1' => 'foo', 'field2' => 'Hello', 'nilfield' => false)), $this->credis->exec());
     }
 
     public function testFalsey()
@@ -310,6 +320,14 @@ class CredisTest extends CredisTestCommon
         $this->credis->pipeline();
         $this->pipelineTestInternal();
         $this->assertEquals(array(), $this->credis->pipeline()->exec());
+
+        try {
+            $iterator = null;
+            $reply = $this->credis->pipeline()->scan($iterator, '*')->exec();
+            $this->assertFalse($reply);
+        } catch (CredisException $e) {
+            $this->assertStringStartsWith('multi()/pipeline() mode can not be used with', $e->getMessage());
+        }
     }
 
     public function testPipelineMulti()
@@ -320,6 +338,14 @@ class CredisTest extends CredisTestCommon
         $this->credis->pipeline()->multi();
         $this->pipelineTestInternal();
         $this->assertEquals(array(), $this->credis->pipeline()->multi()->exec());
+
+        try {
+            $iterator = null;
+            $reply = $this->credis->pipeline()->multi()->scan($iterator, '*')->exec();
+            $this->assertFalse($reply);
+        } catch (CredisException $e) {
+            $this->assertStringStartsWith('multi()/pipeline() mode can not be used with', $e->getMessage());
+        }
     }
 
     public function testWatchMultiUnwatch()
@@ -327,15 +353,16 @@ class CredisTest extends CredisTestCommon
         $this->assertTrue($this->credis->watch('foo', 'bar'));
 
         $reply = $this->credis->pipeline()
-                              ->multi()
-                              ->set('foo', 1)
-                              ->set('bar', 1)
-                              ->exec();
+            ->multi()
+            ->set('foo', 1)
+            ->set('bar', 1)
+            ->exec();
         $this->assertEquals(
             array(
                 true,
                 true,
-            ), $reply
+            ),
+            $reply
         );
         $this->assertTrue($this->credis->unwatch());
     }
@@ -346,6 +373,7 @@ class CredisTest extends CredisTestCommon
         $reply = $this->credis
             ->set('a', 123)
             ->get('a')
+            ->mGet([]) // will be dropped from return result by phpredis
             ->sAdd('b', 123)
             ->sMembers('b')
             ->set('empty', '')
@@ -373,6 +401,7 @@ class CredisTest extends CredisTestCommon
             array(
                 true,               // set('a', 123)
                 '123',              // get('a')
+                //[],                 // get([]) - phpredis doesn't return this
                 1,                  // sAdd('b', 123)
                 array(123),         // sMembers('b')
                 true,               // set('empty', '')
@@ -425,37 +454,73 @@ class CredisTest extends CredisTestCommon
                     'member3' => 3.0,
                     'member2' => 2.0,
                 ),
-            ), $reply
+            ),
+            $reply
         );
     }
 
     public function testTransaction()
     {
         $reply = $this->credis->multi()
-                ->incr('foo')
-                ->incr('bar')
-                ->exec();
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
         $this->assertEquals(array(1,1), $reply);
 
         $reply = $this->credis->pipeline()->multi()
-                ->incr('foo')
-                ->incr('bar')
-                ->exec();
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
         $this->assertEquals(array(2,2), $reply);
 
-        $reply = $this->credis->multi()->pipeline()
-                ->incr('foo')
-                ->incr('bar')
-                ->exec();
-        $this->assertEquals(array(3,3), $reply);
+        $reply = $this->credis->pipeline()
+            ->incr('foo')
+            ->multi()
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
+        $this->assertEquals(array(3,4,3), $reply);
 
         $reply = $this->credis->multi()
-                ->set('a', 3)
-                ->lpop('a')
-                ->exec();
+            ->incr('foo')
+            ->pipeline()
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
+        $this->assertEquals(array(5,6,4), $reply);
+
+        $reply = $this->credis->multi()->pipeline()
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
+        $this->assertEquals(array(7,5), $reply);
+
+        $reply = $this->credis->multi()
+            ->incr('foo')
+            ->pipeline()
+            ->incr('foo')
+            ->incr('bar')
+            ->exec();
+        $this->assertEquals(array(8,9,6), $reply);
+
+        $reply = $this->credis->multi()
+            ->set('a', 3)
+            ->lpop('a') // bad operation for "a"'s type
+            ->exec();
         $this->assertEquals(2, count($reply));
-        $this->assertEquals(TRUE, $reply[0]);
+        $this->assertEquals(true, $reply[0]);
         $this->assertFalse($reply[1]);
+
+        $reply = $this->credis->multi()->pipeline()
+            ->set('a', 3)
+            ->lpop('a') // bad operation for "a"'s type
+            ->exec();
+        $this->assertEquals(2, count($reply));
+        $this->assertEquals(true, $reply[0]);
+        $this->assertFalse($reply[1]);
+
+        $this->assertEquals(array(), $this->credis->multi()->multi()->exec());
+        $this->assertEquals(array(), $this->credis->pipeline()->multi()->multi()->exec());
     }
 
     public function testServer()
@@ -471,7 +536,7 @@ class CredisTest extends CredisTestCommon
         $this->assertEquals('09d3822de862f46d784e6a36848b4f0736dda47a', $this->credis->script('load', 'return 3'));
         $this->assertEquals(3, $this->credis->evalSha('09d3822de862f46d784e6a36848b4f0736dda47a'));
 
-        $this->credis->set('foo','FOO');
+        $this->credis->set('foo', 'FOO');
         $this->assertEquals('FOOBAR', $this->credis->eval("return redis.call('get', KEYS[1])..ARGV[1]", 'foo', 'BAR'));
 
         $this->assertEquals(array(1,2,'three'), $this->credis->eval("return {1,2,'three'}"));
@@ -532,262 +597,369 @@ class CredisTest extends CredisTestCommon
             }
         }
     }
-  public function testDb()
-  {
-      $this->tearDown();
-      $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 1);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertTrue($this->credis->set('database',1));
-      $this->credis->close();
-      $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 0);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertFalse($this->credis->get('database'));
-      $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 1);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals(1,$this->credis->get('database'));
-  }
+    public function testDb()
+    {
+        $this->tearDown();
+        $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 1);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertTrue($this->credis->set('database', 1));
+        $this->credis->close();
+        $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 0);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertFalse($this->credis->get('database'));
+        $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], $this->redisConfig[0]['timeout'], false, 1);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertEquals(1, $this->credis->get('database'));
+    }
 
-  /**
-   * @group Auth
-   */
-  public function testPassword()
-  {
-      $this->tearDown();
-      $this->assertArrayHasKey('password',$this->redisConfig[4]);
-      $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0, $this->redisConfig[4]['password']);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertInstanceOf('Credis_Client',$this->credis->connect());
-      $this->assertTrue($this->credis->set('key','value'));
-      $this->credis->close();
-      $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0, 'wrongpassword');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      try
-      {
-          $this->credis->connect();
-          $this->fail('connect should fail with wrong password');
-      }
-      catch(CredisException $e)
-      {
-          $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
-          $this->credis->close();
-      }
-      $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      try
-      {
-          $this->credis->set('key', 'value');
-      }
-      catch(CredisException $e)
-      {
-          $this->assertStringStartsWith('NOAUTH Authentication required', $e->getMessage());
+    /**
+     * @group Auth
+     */
+    public function testPassword()
+    {
+        $this->tearDown();
+        $this->assertArrayHasKey('password', $this->redisConfig[4]);
+        $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0, $this->redisConfig[4]['password']);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertInstanceOf('Credis_Client', $this->credis->connect());
+        $this->assertTrue($this->credis->set('key', 'value'));
+        $this->credis->close();
+        $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0, 'wrongpassword');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        try {
+            $this->credis->connect();
+            $this->fail('connect should fail with wrong password');
+        } catch(CredisException $e) {
+            if (strpos($e->getMessage(), 'username') !== false) {
+                $this->assertStringStartsWith('WRONGPASS invalid username-password pair', $e->getMessage());
+            } else {
+                $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
+            }
 
-      }
-      try
-      {
-          $this->credis->auth('anotherwrongpassword');
-      }
-      catch(CredisException $e)
-      {
-          $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
-      }
-      $this->assertTrue($this->credis->auth('thepassword'));
-      $this->assertTrue($this->credis->set('key','value'));
-  }
+            $this->credis->close();
+        }
+        $this->credis = new Credis_Client($this->redisConfig[4]['host'], $this->redisConfig[4]['port'], $this->redisConfig[4]['timeout'], false, 0);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        try {
+            $this->credis->set('key', 'value');
+        } catch(CredisException $e) {
+            $this->assertStringStartsWith('NOAUTH Authentication required', $e->getMessage());
+        }
+        try {
+            $this->credis->auth('anotherwrongpassword');
+        } catch(CredisException $e) {
+            if (strpos($e->getMessage(), 'username') !== false) {
+                $this->assertStringStartsWith('WRONGPASS invalid username-password pair', $e->getMessage());
+            } else {
+                $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
+            }
+        }
+        $this->assertTrue($this->credis->auth('thepassword'));
+        $this->assertTrue($this->credis->set('key', 'value'));
+    }
 
-  public function testGettersAndSetters()
-  {
-      $this->assertEquals($this->credis->getHost(),$this->redisConfig[0]['host']);
-      $this->assertEquals($this->credis->getPort(),$this->redisConfig[0]['port']);
-      $this->assertEquals($this->credis->getSelectedDb(),0);
-      $this->assertTrue($this->credis->select(2));
-      $this->assertEquals($this->credis->getSelectedDb(),2);
-      $this->assertTrue($this->credis->isConnected());
-      $this->credis->close();
-      $this->assertFalse($this->credis->isConnected());
-      $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], null, 'persistenceId');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals('persistenceId',$this->credis->getPersistence());
-      $this->credis = new Credis_Client('localhost', 12345);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->credis->setMaxConnectRetries(1);
-      $this->setExpectedExceptionShim('CredisException','Connection to Redis localhost:12345 failed after 2 failures.');
-      $this->credis->connect();
-  }
+    /**
+     * @group Auth
+     */
+    public function testUsernameAndPassword()
+    {
+        $this->tearDown();
+        $this->assertArrayHasKey('password', $this->redisConfig[7]);
+        $this->credis = new Credis_Client($this->redisConfig[7]['host'], $this->redisConfig[7]['port'], $this->redisConfig[7]['timeout'], false, 0, $this->redisConfig[7]['password'], $this->redisConfig[7]['username']);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertInstanceOf('Credis_Client', $this->credis->connect());
+        $this->assertTrue($this->credis->set('key', 'value'));
+        $this->credis->close();
+        $this->credis = new Credis_Client($this->redisConfig[7]['host'], $this->redisConfig[7]['port'], $this->redisConfig[7]['timeout'], false, 0, 'wrongpassword', $this->redisConfig[7]['username']);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        try {
+            $this->credis->connect();
+            $this->fail('connect should fail with wrong password');
+        } catch(CredisException $e) {
+            if (strpos($e->getMessage(), 'username') !== false) {
+                $this->assertStringStartsWith('WRONGPASS invalid username-password pair', $e->getMessage());
+            } else {
+                $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
+            }
 
-  public function testConnectionStrings()
-  {
-      $this->credis->close();
-      $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port']);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals($this->credis->getHost(),$this->redisConfig[0]['host']);
-      $this->assertEquals($this->credis->getPort(),$this->redisConfig[0]['port']);
-      $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host']);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals($this->credis->getPort(),$this->redisConfig[0]['port']);
-      $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port'] . '/abc123');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals('abc123',$this->credis->getPersistence());
-  }
+            $this->credis->close();
+        }
+        $this->credis = new Credis_Client($this->redisConfig[7]['host'], $this->redisConfig[7]['port'], $this->redisConfig[7]['timeout'], false, 0);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        try {
+            $this->credis->set('key', 'value');
+        } catch(CredisException $e) {
+            $this->assertStringStartsWith('NOAUTH Authentication required', $e->getMessage());
+        }
+        try {
+            $this->credis->auth('anotherwrongpassword');
+        } catch(CredisException $e) {
+            if (strpos($e->getMessage(), 'username') !== false) {
+                $this->assertStringStartsWith('WRONGPASS invalid username-password pair', $e->getMessage());
+            } else {
+                $this->assertStringStartsWith('ERR invalid password', $e->getMessage());
+            }
+        }
+        $this->assertTrue($this->credis->auth('thepassword'));
+        $this->assertTrue($this->credis->set('key', 'value'));
+    }
 
-  public function testConnectionStringsTls()
-  {
-      $this->credis->close();
-      $this->credis = new Credis_Client('tls://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port']);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals($this->credis->getHost(),$this->redisConfig[0]['host']);
-      $this->assertEquals($this->credis->getPort(),$this->redisConfig[0]['port']);
-      $this->credis = new Credis_Client('tls://'.$this->redisConfig[0]['host']);
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals($this->credis->getPort(),$this->redisConfig[0]['port']);
-      $this->credis = new Credis_Client('tls://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port'] . '/abc123');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->assertEquals('abc123',$this->credis->getPersistence());
-  }
+    public function testGettersAndSetters()
+    {
+        $this->assertEquals($this->credis->getHost(), $this->redisConfig[0]['host']);
+        $this->assertEquals($this->credis->getPort(), $this->redisConfig[0]['port']);
+        $this->assertEquals($this->credis->getSelectedDb(), 0);
+        $this->assertTrue($this->credis->select(2));
+        $this->assertEquals($this->credis->getSelectedDb(), 2);
+        $this->assertTrue($this->credis->isConnected());
+        $this->credis->close();
+        $this->assertFalse($this->credis->isConnected());
+        $this->credis = new Credis_Client($this->redisConfig[0]['host'], $this->redisConfig[0]['port'], null, 'persistenceId');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->assertEquals('persistenceId', $this->credis->getPersistence());
+        $this->credis = new Credis_Client('localhost', 12345);
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->credis->setMaxConnectRetries(1);
+        if ($this->useStandalone) {
+            $this->setExpectedExceptionShim('CredisException', 'Connection to Redis standalone tcp://localhost:12345 failed after 2 failures.');
+        } else {
+            $this->setExpectedExceptionShim('CredisException', 'Connection to Redis tcp://localhost:12345 failed after 2 failures.');
+        }
+        $this->credis->connect();
+    }
 
-  /**
-   * @group UnixSocket
-   */
-  public function testConnectionStringsSocket()
-  {
-      $this->credis = new Credis_Client(realpath(__DIR__).'/redis.sock',0,null,'persistent');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-      $this->credis->connect();
-      $this->credis->set('key','value');
-      $this->assertEquals('value',$this->credis->get('key'));
-  }
+    public function testConnectionStrings()
+    {
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port']);
+        $this->assertEquals($this->credis->getHost(), $this->redisConfig[0]['host']);
+        $this->assertEquals($this->credis->getPort(), $this->redisConfig[0]['port']);
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host']);
+        $this->assertEquals($this->credis->getPort(), 6379);
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port'] . '/abc123');
+        $this->assertEquals($this->credis->getPersistence(), 'abc123');
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'], 6380);
+        $this->assertEquals($this->credis->getPort(), 6380);
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'], null, null, "abc123");
+        $this->assertEquals($this->credis->getPersistence(), 'abc123');
+    }
 
-  public function testInvalidTcpConnectionString()
-  {
-      $this->credis->close();
-      $this->setExpectedExceptionShim('CredisException','Invalid host format; expected tcp://host[:port][/persistence_identifier]');
-      $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':abc');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-  }
+    public function testConnectionStringsTls()
+    {
+        foreach (['ssl','tls','tlsv1.0','tlsv1.1','tlsv1.2'] as $prefix) {
+            $this->credis = new Credis_Client($prefix.'://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port']);
+            $this->assertEquals($this->credis->getHost(), $this->redisConfig[0]['host']);
+            $this->assertEquals($this->credis->getPort(), $this->redisConfig[0]['port']);
+            $this->assertTrue($this->credis->isTls());
+            $this->credis = new Credis_Client($prefix.'://'.$this->redisConfig[0]['host']);
+            $this->assertEquals($this->credis->getPort(), 6379);
+            $this->assertTrue($this->credis->isTls());
+            $this->credis = new Credis_Client($prefix.'://'.$this->redisConfig[0]['host'] . ':' . $this->redisConfig[0]['port'] . '/abc123');
+            $this->assertEquals($this->credis->getPersistence(), 'abc123');
+            $this->assertTrue($this->credis->isTls());
+            $this->credis = new Credis_Client($prefix.'://'.$this->redisConfig[0]['host'], 6380);
+            $this->assertEquals($this->credis->getPort(), 6380);
+            $this->assertTrue($this->credis->isTls());
+            $this->credis = new Credis_Client($prefix.'://'.$this->redisConfig[0]['host'], null, null, "abc123");
+            $this->assertEquals($this->credis->getPersistence(), 'abc123');
+            $this->assertTrue($this->credis->isTls());
+        }
+    }
 
-  public function testInvalidTlsConnectionString()
-  {
-      $this->credis->close();
-      $this->setExpectedExceptionShim('CredisException','Invalid host format; expected tls://host[:port][/persistence_identifier]');
-      $this->credis = new Credis_Client('tls://'.$this->redisConfig[0]['host'] . ':abc');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-  }
+    public function testTLSConnection()
+    {
+        /**
+        https://www.php.net/manual/en/context.ssl.php
+        > Oddly, if "tls://" is set below, then TLSv1 is forced, but using "ssl://" allows TLSv1.2
 
-  public function testInvalidUnixSocketConnectionString()
-  {
-      $this->credis->close();
-      $this->setExpectedExceptionShim('CredisException','Invalid unix socket format; expected unix:///path/to/redis.sock');
-      $this->credis = new Credis_Client('unix://path/to/redis.sock');
-      if ($this->useStandalone) {
-          $this->credis->forceStandalone();
-      }
-  }
+        and
+        >Recommended use "ssl://" transport.
+        > in php 5.5 ~ 7.1
+        > ssl:// transport = ssl_v2|ssl_v3|tls_v1.0|tls_v1.1|tls_v1.2
+        > tls:// transport = tls_v1.0
+        > after 7.2 ssl:// and tls:// transports is same
+        > php 7.2 ~ 7.3 = tls_v1.0|tls_v1.1|tls_v1.2
+        > php 7.4 ~ 8.1 = tls_v1.0|tls_v1.1|tls_v1.2|tls_v1.3
+         */
 
-  public function testForceStandAloneAfterEstablishedConnection()
-  {
-      $this->credis->connect();
-      if ( ! $this->useStandalone) {
-          $this->setExpectedExceptionShim('CredisException','Cannot force Credis_Client to use standalone PHP driver after a connection has already been established.');
-      }
-      $this->credis->forceStandalone();
-      $this->assertTrue(true);
-  }
-  public function testHscan()
-  {
-      $this->credis->hmset('hash',array('name' => 'Jack','age' =>33));
-      $iterator = null;
-      $result = $this->credis->hscan($iterator,'hash','n*',10);
-      $this->assertEquals($iterator,0);
-      $this->assertEquals($result,['name'=>'Jack']);
-	}
+        $this->credis = new Credis_Client('ssl://'.$this->redisConfig[8]['host'] . ':' . $this->redisConfig[8]['port']);
+        $this->credis->setTlsOptions((array)$this->redisConfig[8]['ssl']);
+        $this->credis->connect();
+        $this->assertTrue(true);
+    }
+
+    /**
+     * @group UnixSocket
+     */
+    public function testConnectionStringsSocket()
+    {
+        $this->credis = new Credis_Client(realpath(__DIR__).'/redis.sock');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+        $this->credis->connect();
+        $this->assertTrue($this->credis->isConnected());
+        $this->credis->set('key', 'value');
+        $this->assertEquals('value', $this->credis->get('key'));
+    }
+
+    public function testInvalidTcpConnectionString()
+    {
+        $this->credis->close();
+        $this->setExpectedExceptionShim('CredisException', 'Invalid host format; expected tcp://host[:port][/persistence_identifier]');
+        $this->credis = new Credis_Client('tcp://'.$this->redisConfig[0]['host'] . ':abc');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+    }
+
+    public function testInvalidTlsConnectionString()
+    {
+        $this->credis->close();
+        $this->setExpectedExceptionShim('CredisException', 'Invalid host format; expected tls://host[:port][/persistence_identifier]');
+        $this->credis = new Credis_Client('tls://'.$this->redisConfig[0]['host'] . ':abc');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+    }
+
+    /**
+     * @group UnixSocket
+     */
+    public function testInvalidUnixSocketConnectionString()
+    {
+        $this->credis->close();
+        $this->setExpectedExceptionShim('CredisException', 'Invalid unix socket format; expected unix:///path/to/redis.sock');
+        $this->credis = new Credis_Client('unix://path/to/redis.sock');
+        if ($this->useStandalone) {
+            $this->credis->forceStandalone();
+        }
+    }
+
+    public function testForceStandAloneAfterEstablishedConnection()
+    {
+        $this->credis->connect();
+        if (! $this->useStandalone) {
+            $this->setExpectedExceptionShim('CredisException', 'Cannot force Credis_Client to use standalone PHP driver after a connection has already been established.');
+        }
+        $this->credis->forceStandalone();
+        $this->assertTrue(true);
+    }
+    public function testHscan()
+    {
+        $this->credis->hmset('hash', array('name' => 'Jack','age' => 33));
+        $iterator = null;
+        $result = $this->credis->hscan($iterator, 'hash', 'n*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, ['name' => 'Jack']);
+    }
+
+    public function testHscanEmptyIterator()
+    {
+        $this->credis->hmset('set', ['foo' => 'bar']);
+        $iterator = 0;
+        $result = $this->credis->zscan($iterator, 'hash', '*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, false);
+    }
+
     public function testSscan()
     {
-        $this->credis->sadd('set','name','Jack');
-        $this->credis->sadd('set','age','33');
+        $this->credis->sadd('set', 'name', 'Jack');
+        $this->credis->sadd('set', 'age', '33');
         $iterator = null;
-        $result = $this->credis->sscan($iterator,'set','n*',10);
-        $this->assertEquals($iterator,0);
-        $this->assertEquals($result,[0=>'name']);
+        $result = $this->credis->sscan($iterator, 'set', 'n*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, [0 => 'name']);
     }
+
+    public function testSscanEmptyIterator()
+    {
+        $this->credis->sadd('set', 'foo', 'bar');
+        $iterator = 0;
+        $result = $this->credis->zscan($iterator, 'set', '*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, false);
+    }
+
     public function testZscan()
     {
-        $this->credis->zadd('sortedset',0,'name');
-        $this->credis->zadd('sortedset',1,'age');
+        $this->credis->zadd('sortedset', 0, 'name');
+        $this->credis->zadd('sortedset', 1, 'age');
         $iterator = null;
-        $result = $this->credis->zscan($iterator,'sortedset','n*',10);
-        $this->assertEquals($iterator,0);
-        $this->assertEquals($result,['name'=>'0']);
+        $result = $this->credis->zscan($iterator, 'sortedset', 'n*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, ['name' => '0']);
     }
+
+    public function testZscanEmptyIterator()
+    {
+        $this->credis->zadd('sortedset', 0, 'name');
+        $iterator = 0;
+        $result = $this->credis->zscan($iterator, 'sortedset', '*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, false);
+    }
+
     public function testscan()
     {
         $seen = array();
-        for($i = 0; $i < 100; $i++)
-        {
+        for ($i = 0; $i < 100; $i++) {
             $this->credis->set('name.' . $i, 'Jack');
             $this->credis->set('age.' . $i, '33');
         }
         $iterator = null;
-        do
-        {
+        do {
             $result = $this->credis->scan($iterator, 'n*', 10);
-            if ($result === false)
-            {
+            if ($result === false) {
                 $this->assertEquals($iterator, 0);
                 break;
-            }
-            else
-            {
-                foreach($result as $key)
-                {
+            } else {
+                foreach ($result as $key) {
                     $seen[$key] = true;
                 }
             }
-        }
-        while($iterator);
+        } while ($iterator);
         $this->assertEquals(count($seen), 100);
     }
 
-  public function testPing()
-  {
-    $pong = $this->credis->ping();
-    $this->assertEquals("PONG",$pong);
-    if (version_compare(phpversion('redis'), '5.0.0', '>='))
+    public function testscanEmptyIterator()
     {
-      $pong = $this->credis->ping("test");
-      $this->assertEquals("test", $pong);
+        $this->credis->set('foo', 'bar');
+        $iterator = 0;
+        $result = $this->credis->scan($iterator, '*', 10);
+        $this->assertEquals($iterator, 0);
+        $this->assertEquals($result, false);
     }
-  }
+
+    public function testPing()
+    {
+        $pong = $this->credis->ping();
+        $this->assertEquals("PONG", $pong);
+        if (version_compare(phpversion('redis'), '5.0.0', '>=')) {
+            $pong = $this->credis->ping("test");
+            $this->assertEquals("test", $pong);
+        }
+    }
 }
