@@ -1,36 +1,40 @@
 <?php
+
 // backward compatibility (https://stackoverflow.com/a/42828632/187780)
 if (!class_exists('\PHPUnit\Framework\TestCase') && class_exists('\PHPUnit_Framework_TestCase')) {
     class_alias('\PHPUnit_Framework_TestCase', '\PHPUnit\Framework\TestCase');
 }
 
-class CredisTestCommon extends \PHPUnit\Framework\TestCase
+if (version_compare(phpversion(), '8.0.0', '>=')) {
+    include 'phpunit-shims/php8.php';
+} else {
+    include 'phpunit-shims/php7.php';
+}
+
+class CredisTestCommon extends CredisTestCommonShim
 {
     protected $useStandalone = false;
     protected $redisConfig = null;
     protected $slaveConfig = null;
 
-    protected function setUp()
+    protected function setUpInternal()
     {
-        if ($this->redisConfig === null)
-        {
+        if ($this->redisConfig === null) {
             $configFile = dirname(__FILE__) . '/redis_config.json';
-            if (!file_exists($configFile) || !($config = file_get_contents($configFile)))
-            {
+            if (!file_exists($configFile) || !($config = file_get_contents($configFile))) {
                 $this->markTestSkipped('Could not load ' . $configFile);
 
                 return;
             }
             $this->redisConfig = json_decode($config);
             $arrayConfig = array();
-            foreach ($this->redisConfig as $config)
-            {
+            foreach ($this->redisConfig as $config) {
                 $arrayConfig[] = (array)$config;
             }
             $this->redisConfig = $arrayConfig;
         }
 
-        if(!$this->useStandalone && !extension_loaded('redis')) {
+        if (!$this->useStandalone && !extension_loaded('redis')) {
             $this->fail('The Redis extension is not loaded.');
         }
     }
@@ -42,18 +46,14 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
      */
     protected function waitForSlaveReplication()
     {
-        if ($this->slaveConfig === null)
-        {
-            foreach ($this->redisConfig as $config)
-            {
-                if ($config['alias'] === 'slave')
-                {
+        if ($this->slaveConfig === null) {
+            foreach ($this->redisConfig as $config) {
+                if ($config['alias'] === 'slave') {
                     $this->slaveConfig = $config;
                     break;
                 }
             }
-            if ($this->slaveConfig === null)
-            {
+            if ($this->slaveConfig === null) {
                 $this->markTestSkipped('Could not load slave config');
 
                 return false;
@@ -67,25 +67,20 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
 
         $start = microtime(true);
         $timeout = $start + 60;
-        while (microtime(true) < $timeout)
-        {
+        while (microtime(true) < $timeout) {
             usleep(100);
             $role = $slaveConfig->role();
-            if ($role[0] !== 'slave')
-            {
+            if ($role[0] !== 'slave') {
                 $this->markTestSkipped('slave config does not points to a slave');
                 return false;
             }
-            if ($role[3] === 'connected')
-            {
+            if ($role[3] === 'connected') {
                 $masterRole = $masterConfig->role();
-                if ($masterRole[0] !== 'master')
-                {
+                if ($masterRole[0] !== 'master') {
                     $this->markTestSkipped('master config does not points to a master');
                     return false;
                 }
-                if ($role[4] >= $masterRole[1])
-                {
+                if ($role[4] >= $masterRole[1]) {
                     return true;
                 }
             }
@@ -95,9 +90,9 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
         return false;
     }
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClassInternal()
     {
-        if(preg_match('/^WIN/',strtoupper(PHP_OS))){
+        if (preg_match('/^WIN/', strtoupper(PHP_OS))) {
             echo "Unit tests will not work automatically on Windows. Please setup all Redis instances manually:".PHP_EOL;
             echo "\tredis-server redis-master.conf".PHP_EOL;
             echo "\tredis-server redis-slave.conf".PHP_EOL;
@@ -107,35 +102,40 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
             echo "\tredis-server redis-auth.conf".PHP_EOL;
             echo "\tredis-server redis-socket.conf".PHP_EOL.PHP_EOL;
         } else {
+            chdir(__DIR__.'/../');
+            if (!file_exists('./tests/tls/ca.crt') || !file_exists('./tests/tls/server.crt')) {
+                // generate SSL keys
+                system('./tests/gen-test-certs.sh');
+            }
             chdir(__DIR__);
             $directoryIterator = new DirectoryIterator(__DIR__);
-            foreach($directoryIterator as $item){
-                if(!$item->isfile() || !preg_match('/^redis\-(.+)\.conf$/',$item->getFilename()) || $item->getFilename() == 'redis-sentinel.conf'){
+            foreach ($directoryIterator as $item) {
+                if (!$item->isfile() || !preg_match('/^redis\-(.+)\.conf$/', $item->getFilename()) || $item->getFilename() == 'redis-sentinel.conf') {
                     continue;
                 }
                 exec('redis-server '.$item->getFilename());
             }
-            copy('redis-master.conf','redis-master.conf.bak');
-            copy('redis-slave.conf','redis-slave.conf.bak');
+            copy('redis-master.conf', 'redis-master.conf.bak');
+            copy('redis-slave.conf', 'redis-slave.conf.bak');
             // wait for redis instances to initialize
             sleep(1);
         }
     }
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClassInternal()
     {
-        if(preg_match('/^WIN/',strtoupper(PHP_OS))){
+        if (preg_match('/^WIN/', strtoupper(PHP_OS))) {
             echo "Please kill all Redis instances manually:".PHP_EOL;
         } else {
             chdir(__DIR__);
             $directoryIterator = new DirectoryIterator(__DIR__);
-            foreach($directoryIterator as $item){
-                if(!$item->isfile() || !preg_match('/^redis\-(.+)\.pid$/',$item->getFilename())){
+            foreach ($directoryIterator as $item) {
+                if (!$item->isfile() || !preg_match('/^redis\-(.+)\.pid$/', $item->getFilename())) {
                     continue;
                 }
                 $pid = trim(file_get_contents($item->getFilename()));
-                if(function_exists('posix_kill')){
-                    posix_kill($pid,15);
+                if (function_exists('posix_kill')) {
+                    posix_kill($pid, 15);
                 } else {
                     exec('kill '.$pid);
                 }
@@ -144,25 +144,8 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
             @unlink('dump.rdb');
             @unlink('redis-master.conf');
             @unlink('redis-slave.conf');
-            @copy('redis-master.conf.bak','redis-master.conf');
-            @copy('redis-slave.conf.bak','redis-slave.conf');
-        }
-    }
-
-
-    //
-    /**
-     * php 7.2 compat fix, as directly polyfilling for older PHPUnit causes a function signature compatibility issue
-     * This is due to the defined return type
-     *
-     * Polyfill for older PHPUnit
-     */
-    protected function createMockShim($originalClassName)
-    {
-        if (method_exists($this, 'getMock')) {
-            return $this->getMock($originalClassName);
-        } else {
-            return parent::createMock($originalClassName);
+            @copy('redis-master.conf.bak', 'redis-master.conf');
+            @copy('redis-slave.conf.bak', 'redis-slave.conf');
         }
     }
 
@@ -170,7 +153,7 @@ class CredisTestCommon extends \PHPUnit\Framework\TestCase
      * php 7.2 compat fix, as directly polyfilling for older PHPUnit causes a function signature compatibility issue
      * This is due to the defined return type
      */
-    public function setExpectedExceptionShim($class, $message = NULL, $code = NULL)
+    public function setExpectedExceptionShim($class, $message = null, $code = null)
     {
         if (method_exists($this, 'setExpectedException')) {
             $this->setExpectedException($class, $message, $code);
